C/R: Fix usage of type long for the appid in commands checkpoint and restart
Port of SVN trunk r5227

The application identifier is a long that is initialized to the pid of the first
process creating the application. The process identifier is of type pid_t, that
is a int. Thus, *currently*, a int is always big enough to store the application
identifier.

Anyway, the way we choose application identifier may change, so it is
safer to fix the type in checkpoint/restart commands.

+++ tools/restart.c
@@ -80,7 +80,7 @@ void parse_args(int argc, char *argv[])
 	if (argc - optind < 2)
 		goto err;
 
-	appid = atoi(argv[argc-2]);
+	appid = atol(argv[argc-2]);
 	version = atoi(argv[argc-1]);
 
 	return ;
